+++ app/src/main/java/de/netid/mobile/sdk/example/MainActivity.kt
@@ -42,6 +42,12 @@ class MainActivity : AppCompatActivity(), NetIdServiceListener, AdapterView.OnIt
         private const val redirectUri = "https://eunid.github.io/redirectApp"
         private const val claims = "{\"userinfo\":{\"email\": {\"essential\": true}, \"email_verified\": {\"essential\": true}}}"
 
+        private const val INVALID_ACCESS_TOKEN = "eyJhbGciOiJkaXIiLCJlbmMiOiJBMjU2R0NNIiwia2lkIjoicGxzX2Flc19rZXktMSIsInppcCI6IkRFRiJ9" +
+            "..-fawL2Eo8FwU10fw.e5sv6xY5frmWPh8wUsrZD3HPfYgBtwDK3IXrmCeRiNyca8XIiWvPEMp_2mSBh7JjChExatir6lLttBbOZuekkVesT7BmGITq8g" +
+            "Q5AkNuoV88jKT0CRyFjvfY6CgabvkJ1CYF8Un0IMXCpyJSb9lt4MNc7sr9shsK1SyWV8vK9mRHk34LMCtnzbIEPbQfR6VYasAMqVXHMv7a3oVXHp8soby" +
+            "wpPBXCHtm03dUtpyfLVzwwSg4GnjaJ9Y4TCw1ocxGTc8DtJ6YaGknuixK1wy3ZiU7h6M-r18yBQf99rMUcR96lz42bhbTjfFnHLpszHeJv5bMrWGJEkob" +
+            "dDly13UnXEmzbP1iZjxRfUy-3Xi_.-S_i_CsstgFS2RvO7l0o9w"
+
         // Using default text / icon
         private val permissionLayerConfig = null
         private val loginLayerConfig = null
@@ -64,6 +70,7 @@ class MainActivity : AppCompatActivity(), NetIdServiceListener, AdapterView.OnIt
         setupAuthorizeButton()
         setupUserInfoButton()
         setupPermissionManagementButtons()
+        setupSetAccessTokenButton()
         setupEndSessionButton()
         updateElementsForServiceState()
 
@@ -178,6 +185,12 @@ class MainActivity : AppCompatActivity(), NetIdServiceListener, AdapterView.OnIt
         }
     }
 
+    private fun setupSetAccessTokenButton() {
+        binding.activityMainButtonSetAccessToken.setOnClickListener {
+            NetIdService.setAccessToken(INVALID_ACCESS_TOKEN)
+        }
+    }
+
     private fun setupEndSessionButton() {
         binding.activityMainButtonEndSession.setOnClickListener {
             it.isEnabled = false
@@ -195,14 +208,12 @@ class MainActivity : AppCompatActivity(), NetIdServiceListener, AdapterView.OnIt
     }
 
     private fun updateElementsForServiceState() {
-        val isUninitialized =
-            serviceState == ServiceState.Uninitialized || serviceState == ServiceState.InitializationFailed
-        val isNotAuthorized =
-            serviceState == ServiceState.InitializationSuccessful || serviceState == ServiceState.AuthorizationFailed
-        val isAuthorized =
-            serviceState == ServiceState.AuthorizationSuccessful || serviceState == ServiceState.UserInfoFailed
-                    || serviceState == ServiceState.UserInfoSuccessful || serviceState == ServiceState.PermissionWriteSuccessful || serviceState == ServiceState.PermissionWriteFailed
-                    || serviceState == ServiceState.PermissionReadFailed || serviceState == ServiceState.PermissionReadSuccessful
+        val isUninitialized = serviceState == ServiceState.Uninitialized || serviceState == ServiceState.InitializationFailed
+        val isNotAuthorized = serviceState == ServiceState.InitializationSuccessful || serviceState == ServiceState.AuthorizationFailed
+        val isAuthorized = serviceState == ServiceState.AuthorizationSuccessful || serviceState == ServiceState.UserInfoFailed ||
+            serviceState == ServiceState.UserInfoSuccessful || serviceState == ServiceState.PermissionWriteSuccessful ||
+            serviceState == ServiceState.PermissionWriteFailed || serviceState == ServiceState.PermissionReadFailed ||
+            serviceState == ServiceState.PermissionReadSuccessful
 
         binding.activityMainButtonInitialize.isEnabled = isUninitialized
         binding.activityMainButtonAuthorize.isEnabled = isNotAuthorized
@@ -210,6 +221,7 @@ class MainActivity : AppCompatActivity(), NetIdServiceListener, AdapterView.OnIt
         binding.activityMainButtonEndSession.isEnabled = isAuthorized
         binding.activityMainButtonPermissionWrite.isEnabled = isAuthorized
         binding.activityMainButtonPermissionRead.isEnabled = isAuthorized
+        binding.activityMainButtonSetAccessToken.isEnabled = isAuthorized
         binding.activityMainCheckBoxShippingAddress.isEnabled = isUninitialized
         binding.activityMainCheckBoxBirthdate.isEnabled = isUninitialized